requirements.txt
fastapi
linkheader
python-slugify
tinydb
uvicorn
